+++ packages/common/src/index.ts
@@ -1,3 +1,5 @@
 export * from './monitoring';
 export * from './interceptors';
+export * from './services';
+export * from './modules';
 export * from './controller';
