+++ packages/common/src/modules/temporal.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { TemporalWorkflowService } from '../services/temporal.service';
+
+@Module({
+  providers: [TemporalWorkflowService],
+})
+export class TemporalWorkflowModule {}
