+++ sample/05-temporal-package/src/app.module.ts
@@ -3,7 +3,8 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TemporalModule } from 'nestjs-temporal';
 import * as activities from './temporal/activities';
-import { TemporalWorkflowService } from './temporal.service';
+// import { TemporalWorkflowService } from './temporal.service';
+import { TemporalWorkflowService } from '@samagra-x/stencil';
 
 @Module({
   imports: [
